+++ .idea/workspace.xml
@@ -8,8 +8,17 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="a50ee93e-5af6-4fb6-a1c1-6af945bc26a8" name="변경" comment="한수현 + 서희림">
+      <change afterPath="$PROJECT_DIR$/fastapi/database.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/fastapi/utils.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/fastapi/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/engineering.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/engineering.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/logout.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/logout.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/management.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/management.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/model_deployment.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/model_deployment.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/social.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/social.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/routers/user_management.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/routers/user_management.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/fastapi/superset.py" beforeDir="false" afterPath="$PROJECT_DIR$/fastapi/superset.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -56,17 +65,17 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">{
-  &quot;keyToString&quot;: {
-    &quot;Python.main.executor&quot;: &quot;Run&quot;,
-    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;,
-    &quot;git-widget-placeholder&quot;: &quot;EO3-91-Backend-ModelManagement1&quot;,
-    &quot;last_opened_file_path&quot;: &quot;C:/Users/KFQ/Desktop/git/BACKEND&quot;,
-    &quot;run.code.analysis.last.selected.profile&quot;: &quot;aDefault&quot;,
-    &quot;settings.editor.selected.configurable&quot;: &quot;com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable&quot;
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "Python.main.executor": "Run",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "SHARE_PROJECT_CONFIGURATION_FILES": "true",
+    "git-widget-placeholder": "EO3-103-Backend-Router-Modularization",
+    "last_opened_file_path": "C:/Users/KFQ/Desktop/git/BACKEND",
+    "run.code.analysis.last.selected.profile": "aDefault",
+    "settings.editor.selected.configurable": "com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable"
   }
-}</component>
+}]]></component>
   <component name="Push.Settings">
     <option name="SHOW_DETAILS_PANEL" value="false" />
   </component>
