+++ fastapi/routers/logout.py
@@ -2,6 +2,11 @@ from fastapi import APIRouter
 
 router = APIRouter()
 
-@router.get("/")
-async def management_home():
-    return {"message": "Welcome to User Logout Home"}
+# @router.post("/")
+# async def logout():
+#     """
+#     로그아웃 엔드포인트:
+#     - 클라이언트의 세션을 종료하거나 쿠키를 삭제합니다.
+#     """
+#     response = {"message": "로그아웃되었습니다."}
+#     return response
