+++ fastapi/superset.py
@@ -7,3 +7,4 @@ async def get_superset_data():
         response = await client.get(superset_url)  # 헤더에서 Authorization 제거
         response.raise_for_status()  # 오류 발생 시 예외 발생
         return response.json()  # JSON 데이터 반환
+
